% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise_to_outcomes.R
\name{add_noise_to_outcomes}
\alias{add_noise_to_outcomes}
\title{Add noise to match outcomes}
\usage{
add_noise_to_outcomes(outcome, error_prob = 0.1)
}
\arguments{
\item{outcome}{Original match outcome. Must be one of 0, 0.5, or 1.}

\item{error_prob}{The probability an outcome switches from a 0 to 1 or
vice-versa}
}
\value{
a numeric vector.
}
\description{
This function can be used to flip a set proportion of match outcomes. Useful
for testing models under varying levels of signal to noise ratios.
}
\details{
Input outcomes of 0.5 will be randomy rounded to either 0 or 1.
}
\examples{
# Create some sample outcomes
set.seed(1)
test_df <- data.frame(
   true_outcome = sample(c(0,1), 1000, replace=TRUE)
)

# No noise
test_df$noisy_outcome = add_noise_to_outcomes(test_df$true_outcome, error_prob = 0)
table(test_df)

# Test different noise levels
test_df$noisy_outcome = add_noise_to_outcomes(test_df$true_outcome, error_prob = 0.1)
table(test_df)

test_df$noisy_outcome = add_noise_to_outcomes(test_df$true_outcome, error_prob = 0.5)
table(test_df)

test_df$noisy_outcome = add_noise_to_outcomes(test_df$true_outcome, error_prob = 1)
table(test_df)

}
