% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afl_df}
\alias{afl_df}
\title{AFL match outcomes}
\format{A data frame with four variables:
\describe{
\item{\code{match_index}}{Unique index for the matches.}
\item{\code{home_team}}{Home team.}
\item{\code{away_team}}{Away team.}
\item{\code{outcome}}{Outcome of the match.}
\item{\code{date}}{Date of match.}
}}
\usage{
afl_df
}
\description{
A data.frame giving a history of AFL match outcomes since the start of the
2000 season.
}
\keyword{datasets}
