% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELO.R
\name{ELO}
\alias{ELO}
\title{Calculate ELO ratings}
\usage{
ELO(
  match_data,
  init_data = NULL,
  init_rating = 2200,
  eta = 27,
  p1_advantage = 0,
  save_history = TRUE,
  sort = TRUE
)
}
\arguments{
\item{match_data}{A data frame containing four columns: (1) a numeric
vector denoting the time period in which the game took place (2) a numeric
or character identifier for player one (3) a numeric or character identifier
for player two and (4) the result of the game expressed as a number,
typically equal to one for a player one win, zero for a player two win and
one half for a draw.}

\item{init_data}{Initialise a rating model with ratings component of the
returned list of the output from a previous \code{ELO()} call or a data.frame with
columns Player and Rating.}

\item{init_rating}{The initial rating for players not appearing in
\code{init_data}.}

\item{eta}{The learning rate.}

\item{p1_advantage}{Player 1 advtange parameter. Either a single value or a
vector equal to the number of rows in \code{match_data}.}

\item{save_history}{If \code{TRUE} return the rating history for each player.}

\item{sort}{If \code{TRUE}, sort the output ratings from highest to lowest.}
}
\value{
A list object of class "\code{mELO_rating}" with the following
components:
\describe{
  \item{ratings}{A data frame of the results at the end of the final time period.
  The variables are self explanatory except for Lag, which represents the
  number of time periods since the player last played a game. This is equal
  to zero for players who played in the latest time period, and is also zero
  for players who have not yet played any games.}
  \item{history}{An array containing the ratings history for all
  players.}
  \item{p1_advantage}{A single value or a vector values for the
  advantage Player 1 had.}
  \item{eta}{The learning rate.}
  \item{type}{The type of model. In this case, "ELO".}
  \item{preds}{The player 1 success probabilities predicted prior to adjusting to the
  outcome of the match.}
  \item{outcomes}{The outcome for player 1 for each match.}
  \item{preds_logloss}{The mean logloss error for all predictions.}

}
}
\description{
This function implements the ELO rating system for performing pairwise
comparisons.
}
\examples{
# AFL data
head(afl_df)
afl_ELO_ratings <- ELO(afl_df[1:4], eta = 20, p1_advantage = 70)
afl_ELO_ratings
}
\references{
Elo, Arpad (1978) The Rating of Chessplayers, Past and Present. Arco. ISBN 0-668-04721-6.
}
