% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_C_given_A.R
\name{find_C_given_A}
\alias{find_C_given_A}
\title{Estimate C matrix from an Advantage/Disadvantage matrix}
\usage{
find_C_given_A(A_mat, k = 1)
}
\arguments{
\item{A_mat}{An advantage/disadvantage matrix.}

\item{k}{The number of columns of the C matrix. More complex
non-transitive interactions require higher values of k.}
}
\value{
A list with three components:
\describe{
  \item{C_mat}{The estimated C matrix}
  \item{error}{The squared reconstruction error}
  \item{est_req_k}{The estimated required value for k to minimise the
  reconstruction error.}
}
}
\description{
Use this function to estimate a possible C matrix from a skew-symmetric A
matrix. Recall that \eqn{A = C \Omega C'}.
}
\examples{
# Generate C mat
set.seed(112123)
m <- 4
k <- 1
C_mat <- matrix(
    runif(m*2*k, 0, 10),
  ncol = 2*k
)


Omega_mat <- construct_omega(k)
A_mat <- C_mat \%*\% Omega_mat \%*\% t(C_mat)

C_mat_list <- find_C_given_A(A_mat)
C_mat_list
}
