% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adv_mat.R
\name{get_adv_mat}
\alias{get_adv_mat}
\title{Get advantage matrix from a mELO model}
\usage{
get_adv_mat(model)
}
\arguments{
\item{model}{a \code{mELO_rating} model created with a \code{mELO()} function.}
}
\value{
a matrix
}
\description{
Returns the player to player advantage/disadvantage matrix for a mELO model.
the matrix is calculated as \eqn{C \Omega C'}.
}
\examples{
# rock paper scissors model
rps_model <- mELO(rps_df, k=1)
get_adv_mat(rps_model)

# rock paper scissors fire water model
rpsfw_model <- mELO(rpsfw_df, k=2)
get_adv_mat(rpsfw_model)
}
