% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logloss.R
\name{logloss}
\alias{logloss}
\title{Calculate logloss for evaluating predictions}
\usage{
logloss(prediction, outcome, tol = .Machine$double.neg.eps)
}
\arguments{
\item{prediction}{A vector of estimated probabilities.}

\item{outcome}{A vector of observed outcomes.}

\item{tol}{Numerical tolerance. Can also be used to threshold errors
for really bad predictions, or when you don't want a model to be penalized
too strongly in the presence of high dispersion. Default is
.Machine$double.neg.eps.}
}
\value{
a numeric vector
}
\description{
Calculate logloss or cross-entropy for a set of predictions.
}
\examples{
preds <- c(0.9, 0.1, 0.8, 0.5)
outcomes <- c(1, 0, 1, 0)
logloss(preds, outcomes)

# Thresholding large errors for bad predictions
preds <- c(0.000001)
outcomes <- c(1)
logloss(preds, outcomes)
logloss(preds, outcomes, 0.01)
}
