% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pred_mat.R
\name{model_pred_mat}
\alias{model_pred_mat}
\title{Return pairwise predictions for specified agents}
\usage{
model_pred_mat(model, agents, round = 4)
}
\arguments{
\item{model}{An object of class \code{mELO_rating}.}

\item{agents}{a vector of agents/players/teams you would like to pairwise
predictions for.}

\item{round}{Round results to how many figures. Default is 4.}

\item{actual_matrix}{Should result be an actual matrix? Default is FALSE and
the output is a data.frame.}
}
\value{
a data.frame.
}
\description{
Given a mELO_rating model and a vector of agents, this function gives the
predicted success probabilities for each combination. This is useful for mELO
models as the ratings alone is not useful for predicting outcomes.
}
\examples{
# Fit a mELO model that can handle these types of interactions.
rps_mELO <- mELO(rps_df, k=1)
rps_mELO

# Get pairwise predictions
model_pred_mat(
   rps_mELO,
   unique(rps_df[[2]])
)
}
