% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mELO_rating.R
\name{plot.mELO_rating}
\alias{plot.mELO_rating}
\title{Plot historical player ratings for a mELO_rating object}
\usage{
\method{plot}{mELO_rating}(
  modelled_ratings,
  xlab = "Time Period",
  ylab = paste(modelled_ratings$type, "Ratings"),
  main = NULL
)
}
\arguments{
\item{modelled_ratings}{An object of class \code{mELO_rating}.}

\item{xlab}{Title for x axis.}

\item{ylab}{Title for y axis.}

\item{main}{Title for the plot.}

\item{x_labels}{Custom labels for the x axis.}
}
\value{
a plot
}
\description{
Plot historical player ratings for a \code{mELO_rating} object fit with the
\code{ELO()} or \code{mELO()} functions.
}
\examples{
# History of AFL ratings
head(afl_df)

afl_ELO <- ELO(afl_df[,1:4])
plot(afl_ELO)
}
