% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_c_mat.R
\name{plot_c_mat}
\alias{plot_c_mat}
\title{Plot the evolution of the C matrix for a mELO model}
\usage{
plot_c_mat(mELO_model)
}
\arguments{
\item{mELO_model}{a \code{mELO_rating} object with \code{type = mELO}. The
output of a \code{mELO()} function call.}
}
\value{
a plot.
}
\description{
This function plots the evolution of the elements of the C matrix for a mELO
model. The C matrix has a row vector for each player, and k columns. Where k
determines the complexity of non-transitive interactions being modelled.
}
\examples{
head(rpsfw_df)

# mELO model with k=1
rpsfw_mELO_1 <- mELO(rpsfw_df, k=1)
plot_c_mat(rpsfw_mELO_1)
    # not stable, mELO with k=1 can't accurately model the dynamics

# mELO model with k=2
rpsfw_mELO_2 <- mELO(rpsfw_df, k=2)
plot_c_mat(rpsfw_mELO_2)
    # stable, good estimates
}
