% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mELO_rating.R
\name{predict.mELO_rating}
\alias{predict.mELO_rating}
\title{Predict results of a game from a ELO or mELO model}
\usage{
\method{predict}{mELO_rating}(
  object,
  new_match_data,
  min_games = 15,
  default_ratings = NULL,
  p1_advantage = 0,
  thresh
)
}
\arguments{
\item{object}{An object of class \code{mELO_rating}.}

\item{new_match_data}{A data frame containing four columns: (1) a numeric
vector denoting the time period in which the game took place (2) a numeric
or character identifier for player one (3) a numeric or character identifier
for player two.}

\item{min_games}{A single value. If the number of games of either player is
below this value, the prediction will be based on the \code{default_ratings}
parameter.}

\item{default_ratings}{The rating to be used for agents or players who have
not yet played \code{min_games}.}

\item{p1_advantage}{Player 1 advantage parameter. A single value or numeric
vector with length equal to the number of rows in \code{new_match_data}. Can
be though of as representing first move or home ground advantage.}

\item{thresh}{A single value. If given, a binary vector is returned
indicating whether the prediction is greater than this value.}
}
\value{
A numeric vector of predictions, which may contain missing values.
}
\description{
This function gives predictions of success probabilities for agent or
player 1 from a fitted ELO or mELO model (a \code{mELO_rating} oject).
}
\examples{
# Rock paper scissors
head(rps_df)

# Note that ELO doesn't perform well
rps_ELO <- ELO(rps_df)
rps_ELO
ELO_preds <- predict(
    rps_ELO,
    head(rps_df)
)
cbind(
    head(rps_df),
    ELO_preds
)
    # Predictions are all ~0.5

# Fit a mELO model that can handle these types of interactions.
rps_mELO <- mELO(rps_df, k=1)
rps_mELO
# Inspect advantage matrix
get_adv_mat(rps_mELO)
# Get predictioncs
mELO_preds <- predict(
    rps_mELO,
    head(rps_df)
)
cbind(
    head(rps_df),
    mELO_preds
)
    # Much better predictions!
}
