% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mELO_rating.R
\name{print.mELO_rating}
\alias{print.mELO_rating}
\title{Print simple output of mELO_rating object to screen}
\usage{
\method{print}{mELO_rating}(modelled_ratings, digits = 1)
}
\arguments{
\item{modelled_ratings}{A mELO_rating object.}

\item{digits}{Number of significant digits in the ratings column to be output.}
}
\value{
text to screen
}
\description{
Print simple output of mELO_rating object to screen
}
\examples{
# rock paper scissors model
rps_model <- mELO(rps_df, k=1)
# print.mELO_rating is automatically called:
rps_model
}
