% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rps_df}
\alias{rps_df}
\title{Rock Paper Scissors data}
\format{A data frame with four variables:
\describe{
\item{\code{time_index}}{Time point match was played.}
\item{\code{throw_1}}{Throw 1.}
\item{\code{throw_2}}{Throw 2.}
\item{\code{outcome}}{Outcome of the match.}
}}
\usage{
rps_df
}
\description{
A data.frame of rock-paper-scissors-matches.
}
\keyword{datasets}
