% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpsfw_df}
\alias{rpsfw_df}
\title{Rock Paper Scissors Fire Water data}
\format{A data frame with four variables:
\describe{
\item{\code{time_index}}{Time point match was played.}
\item{\code{throw_1}}{Throw 1.}
\item{\code{throw_2}}{Throw 2.}
\item{\code{outcome}}{Outcome of the match.}
}}
\usage{
rpsfw_df
}
\description{
A data.frame of Rock-Paper-Scissors-Fire-Water matches. This is a variant
of rock-paper-scissors that accommodates 5 different moves but with a more
complex non-transitive behaviour.
}
\keyword{datasets}
