% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpssl_df}
\alias{rpssl_df}
\title{Rock Paper Scissors Spock Lizard data}
\format{A data frame with four variables:
\describe{
\item{\code{time_index}}{Time point match was played.}
\item{\code{throw_1}}{Throw 1.}
\item{\code{throw_2}}{Throw 2.}
\item{\code{outcome}}{Outcome of the match.}
}}
\usage{
rpssl_df
}
\description{
A data.frame of Rock-Paper-Scissors-Spock-Lizard matches. This is a variant
of rock-paper-scissors that accommodates 5 different moves with similar
non-transitive behaviours.
}
\keyword{datasets}
