% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric
}
\description{
A simple function for obtaining \eqn{f(x) = 1/(1+exp(-x))}.
}
\examples{
sigmoid(c(-10, -1, 0, 1, 10))
}
